User defined aperture
User can pass in a shape, radius (for circle), or length x width (for rectangle for an aperture. The user can define whether or not they would like the aperture to follow the pointing model. The user can turn on whatever systematics corrections they would like to take into account.