give TargetData a boolean lite flag
The visualize class has been changed to check if the TargetData object it was fed is loaded from a lite fits file. However, this raises an error now because the lite property is currently only set in the load function. It needs to be set to False by default in the init when the object has full access to everything and visualize should work.
https://github.com/afeinstein20/eleanor/blob/bd5b8cdc6d80bda9ebab2b464f07aab24bc93f7f/eleanor/visualize.py#L132-L135