breaking some components from update.py out
Within update.py, there are several lines that assign a target and the year for that sector. These pieces of information don't need to be buried in update.py. I've taken them out and created two new functions in assignments.py: (1) assign_year assigns the year during which a given sector was observed and (2) assign_target assigns a target that was observed during that sector, which is then used to download some metadata information